Create Hi I am MaDonna McAllisterREADME.md
I have a special interest in creating this website called Makem' Sparkle Glass Beads & Crystal Artwork I make- Currently I've been making this Artwork for 20 + years. This has been only a coping strategy to handle stressors in my life. This also has been a way I haven given back to to those who have taken care of me through out my years alive. Currently struggling with employment sustainability for long- term goals. It is my hope to support myself in the working World. Moreover, as I am seeking to distribute my Artwork as a means and source of income. This is a scary venture for me to take on at this time However this feels is not as important as what I hope to be giving  back, as an attempt to make the World a brighter & colorful place through reflective prismatic illumination.